fix: allow "Times New Roman" substitute font
"Liberation Serif" is a layout compatible font from Liberation Fonts. This means that the characters of "Liberation Serif" are identical in width and height to the characters of "Times New Roman". This is helpful for people who don't use Windows, or don't have Times New Roman available on their system.

More information can be found here:
https://github.com/liberationfonts/liberation-fonts